tets: Split localization module task tests
